[Xamarin.Android.Build.Tasks] Add CancelableTask.
The `ICancelableTask` is implemented by the AsyncTask
so that long background processes can respond to a `Cancel`
requrest from MSBuild.

The problem here is that many of our tasks which we might
want to cancel, do not derive from `AsyncTask`. Given that
allot of the time a task wont be using the features of
`AsyncTask`, it makes sense to introduce a new bass class.

This new class implements `ICancelableTask` and will allow
normal `Tasks` to use those features. This can be done by
overriding the `Cancel` method. Or alternatively using the
`CancellationToken` property `Token` to `Register` callbacks
or simply check `Token.IsCancellationRequested`.

`AsyncTask` has been updated to use the new base class.