[Xamarin.Android.Build.Tasks] Add CancelableTask (#2067)
The `ICancelableTask` is implemented by `AsyncTask` so that long
background processes can respond to `Cancel()` requests from MSBuild.

The problem here is that many of our tasks which we might want to
cancel do not derive from `AsyncTask`.  Given that a lot of the time
a task wont be using the features of `AsyncTask`, it makes sense to
introduce a new `CancelableTask` bass class.

`CancelableTask` implements `ICancelableTask` and will allow "normal"
`Task`s to use those features.  This can be done by overriding the
`Cancel()` method, or alternatively by using the `CancellationToken`
property `Token` to `Register` callbacks, or simply check
`Token.IsCancellationRequested`.

`AsyncTask` has been updated to use the new base class.