[Xamarin.Android.Build.Tasks] shorten $(IntermediateOutputPath) for MAX_PATH (#2084)
* [Xamarin.Android.Build.Tasks] shorten $(IntermediateOutputPath) for MAX_PATH

Context: https://github.com/xamarin/xamarin-android/pull/1628
Fixes: http://work.devdiv.io/667203

Customers are reporting MAX_PATH (or PathTooLongException) related
failures when upgrading Visual Studio from 15.7 to 15.8. This appears
to be due to the change of moving `$(IntermediateOutputPath)` from
`obj\Debug\` to `obj\\Debug\MonoAndroid81`.

Note we have two things we can do here to improve this:
- Rename `MonoAndroid81` to just `81`
- Fix the double slash `\\`

There was also the comment of using `$(AndroidApiLevel)`, but it feels
a bit safer to use what we have currently and shorten it. In the long
run, we should revisit the current approach entirely.

It appears during MSBuild evaluation, the `$(PlatformName)` is empty
(and not equal to `AnyCPU`). So we should just drop this extra
directory entirely, since it is causing a double-slash `\\`.

The original goals of PR #1628 were to fix this scenario:
- Build with `MonoAndroid81`
- Build with `MonoAndroid90` or *something else*, something in `obj`
  is completely broken and requires #deletebinobj

We should fix the above scenario properly down the road, and we won't
need an extra `81` folder at all.

* [tests] fixed failing test on Windows

Updated XamarinAndroidProject to include a `TargetFrameworkAbbreviated` property.