Revert "[Xamarin.Android.Build.Tasks] more fixes for parallel builds (#2063)"
This *partially* reverts commit 73f7f47e15495755bfd1da1448cd7d6e305fd71f.

We want the change to the test, because it was failing on MacOS.

However, I noticed a performance regression in
`ResolveLibraryProjectImports`. It went from ~1 sec to ~3 sec for me.

Since we shouldn't make our build times *worse*, let's revert this
change for now. It only *helped* parallel builds on Mac, but did not
fully fix them.