[Xamarin.Android.Build.Tasks] Fix Missing fields when using aapt2. (#2058)
Fixes #2047

When we implemented the `ManagedResouceParser` the initial thought
was that all the fields we present would be unique. So when we
checked to see if the container class already had a field with
the same name we uses a Case Insensitive comparision to do that.

We a result this broke a number of sample applicaitons that QA have.
Looking at the `ApiDemo` issue.. we have this problem

	Resources/layout/animation_seeking.xml` has a item with id = `android:id="@+id/seekBar"
	Resources/layout/content_browser.xml` has id = `android:id="@+id/seekbar"
	Resources/layout/content_browser_nav.xml` has id = `android:id="@+id/seekbar"

Note the different casing for `seekBar`. With the old `aapt` based
R.java resource parser this results in BOTH `seekbar` and `seekBar`
being available in the `Resource.Desogner.cs` file.

Because of the Case Insensitive check in the new parser these two
fields are merged into one. It would be on a first found basis.
So if `seekBar` was the first field added , then that is exposed.
The other `seebar` field would be ignored.

We might see this as a problem in our parser but aapt and aapt2 both
produce the following in the `R.txt` file

	int id seekBar 0x7f08018e
	int id seekbar 0x7f08018f

So the fix in this case is to just do a normal case check to ensure
that if we have fields with different casing, they are handled
correctly.