Remove all traces of armeabi (ARMv5)
Android NDK r17 dropped support for armeabi and any attempt to build projects
targetting this ABI fails with an error. In preparation for NDK r17 this commit
removes all the traces of armeabi from Xamarin.Android tree.

In order to prevent confusion, all the instances where we used `armeabi` but
meant `armeabi-v7a` have been renamed to the latter form.