Make {x,n}unit runner logging level configurable from command line
The LogWriter class used by both xUnit and NUnit test runners has configurable
minimum logging level which defaults to `Info`. It is not possible to configure
the level from command line, e.g. when starting the instrumentation. This commit
fixes the problem by introducing the `loglevel` parameter which can be sent to
the instrumentation using this syntax:

     -e "loglevel info"

The acceptable log level values are (the strings are case-insensitive):

   * Wtf
   * Error
   * Warning
   * Info
   * Debug
   * Verbose

The default in the logger class is reasonable `Info`, however this commit
changes the value used by the `RunTestApks` target to `Verbose` in order to try
to figure out why the tests are failing on us when run on PR builders.