Fix AOT build
The `leeroy` make target failed with a "file not found" exception while
attempting to invoke the `cross-arm` crosscompiler and the reason for this was
that `cross-arm` was never built because of the `armeabi-v7a` target missing in
the `ALL_AOT_ABIS` make variable in the `build-tools/scripts/BuildEverything.mk`
file.

This commit adds the missing target architecture and fixes the build.

In addition to the above changes this commit modifies APK tests to:

    * target Android SDK 22 (to maintain uniformity across tests and also to
      make sure that the code which determines location of the on-device test
      result files returns the same location for all the tests - it relies on
      APIs introduced in API level 19)
    * request external storage premissions. This ensures that we can create the
      test result files in a location that's publically accessible. If access to
      external storage is denied our test code will create the results file in
      the application data directory and on Android 28 the file will be only
      readable by owner, thus making `adb pull` fail to download it.