[Xamarin.Android.Build.Tasks] fix incremental builds for Xamarin.Forms projects (#2088)
@StephaneDelcroix noticed a build performance problem when building
Xamarin.Forms projects:

  - Build
  - Build again with no changes
  - Slow targets such as `_UpdateAndroidResgen` and
    `_GenerateJavaStubs` run again!

The cause appeared to be this:

	Building target "_UpdateAndroidResgen" completely.
	Input file "obj/Debug/res/layout/tabbar.xml" is newer than output file "obj/Debug/R.cs.flag".

I modified the `CheckTimestamps()` test to replicate this issue:

  - Added Xamarin.Forms NuGet packages
  - Added `Tabbar.axml` from the Xamarin.Forms template

This surfaced several problems with timestamps and Xamarin.Android
building incrementally.

~~ Problem 1 ~~

In the `_GenerateJavaStubs` target, the `<ConvertResourcesCases/>`
MSBuild task is not supplied an `AndroidConversionFlagFile`.  It uses
this file to compare timestamps and decide if XML files need to be
fixed up or not.

Supplying `$(_AndroidResgenFlagFile)` seemed to be the way to go here,
since `_UpdateAndroidResgen` runs right before this task.  It also
solves the original problem from the `Tabbar.axml` file.

I also realized we should be using a "stamp" file for the
`_GenerateJavaStubs` MSBuild target, in general.

This gives us several benefits:

  - `<ConvertResourcesCases/>` has a proper
    `AndroidConversionFlagFile` it can use for checking timestamps
  - The `Outputs` of the target are drastically simplified, which may
    help performance in MSBuild evaluating if the target should run
  - It is likely more precise/correct.

So I added a new file in `$(IntermediateOutputPath)`,
`obj\Debug\_javastubs.stamp` we use to determine if
`_GenerateJavaStubs` should run.

I also tried to do things "the right way", such as:

  - Adding the stamp file to the `@(FileWrites)` item group
  - Made sure the file is deleted during a `Clean`

After these changes, two instances of `<ConvertResourcesCases/>` were
*still* fighting each other.  They would always run and update the
timestamps on one file, causing the other one to run.

`ConvertResourcesCases` runs twice:

  - In the `_UpdateAndroidResgen` target
  - In the `_GenerateJavaStubs` target

But these are both working with the same files, so in the second case,
we need to make sure `_GenerateJavaStubs` updates the timestamp on the
*first* flag file so the target doesn't run every time.

~~ Problem 2 ~~

Timestamps of Xamarin.Forms assemblies in `obj\Debug\linksrc` were
out of date!

I updated the `_CopyIntermediateAssemblies` target to update the
timestamps of all files within this directory.  It did not appear to
be doing it correctly, and was only running `<Touch/>` on a subset of
assemblies.

~~ Problem 3 ~~

After fixing Problem 2, the next problem I noticed were `*.pdb` and
`*.mdb` files in `obj\Debug\linksrc` were out of date!

The problem here was a call to `<Touch />`:

	<Touch Files="@(_DebugFilesCopiedToLinkerSrc)" />

The `@(_DebugFilesCopiedToLinkerSrc)` did not appear to contain any
items!

I updated the `_CopyPdbFiles` and `_CopyMdbFiles` targets to use the
correct item groups.

~~ Problem 4 ~~

Lastly, after fixing the previous three problems, the timestamps of
output assemblies in `obj\Release\android\assets` were not correct
for `Release` builds.

I had made a change for this in the past, which was making `Debug`
builds work correctly:

https://github.com/xamarin/xamarin-android/pull/2028

At the time I wrote #2028, since the test wasn't using Xamarin.Forms,
I did not see the problem occurring in `Release` mode.  It *does*
occur if you are using NuGet packages.

I replicated what I changed in the `_LinkAssembliesNoShrink` target,
and also renamed the item groups to be unique:

  - `_LinkAssembliesNoShrinkFiles`
  - `_LinkAssembliesShrinkFiles`

~~ Other changes ~~

The `CheckTimestamps()` test also was verifying timestamps in
`$(OutputPath)` such as `bin\Debug`.

It appears the core MSBuild targets are leaving timestamps as-is in
the `CopyFilesToOutputDirectory` and `_CopyFilesMarkedCopyLocal`
targets. Assemblies from Xamarin.Forms NuGet are out of date here, but
I don't think we should do anything, since the core MSBuild targets
are doing this.

For now, I removed the checks for `$(OutputPath)` in this test, and
cleaned up the test a bit.

Update `Files.ExtractAll()` so that the timestamps of extracted files
are set to "now", in order to fix some failing unit tests on macOS.