[Mono.Android] make Stream seekable.
There are some Android API that returns java.io.InputStream, which we
convert to System.IO.Stream and support `skip()` method, which works
like `Seek()` in .NET land.
Our Stream implementation is sadly less functional and `Seek()` throws
NotSupportedException. It is particularly problematic when we try to
open and seek Android assets via `AssetManager.Open()`.

Here is an "aggressive" implementation that supports those operations
(Java API for seek operation is fugly and cannot be simply implemented).