[Xamarin.Android.Tools.BootstrapTasks] Make output of UnzipDirectoryChildren less messy
The lines are all messed up because the underlying unzip command uses 2 steps to print a line: first the line content, then the '\n'. This can race across processes and makes the output quite hard to read.

This fixes that by making sure we print the while line+'\n' at the same time. This doesn't reduce the chance of broken line to 0, but it greatly reduces it anyway.