[Xamarin.Android.Build.Tasks] Add ConvertCustomView Task (#2129)
Fixes: https://github.com/xamarin/xamarin-android/issues/2100

Currently the `<ConvertResourcesCases/>` task is called twice on ALL
resources referenced by the project.  This is terribly inefficient,
but unfortunately is required: the first pass is done before the
`Compile` target, and fixes up the casing for things like
drawables/etc. so that they're ready to be processed by `aapt`.
The second pass is done after `<GenerateJavaStubs/>`, which happens
*after* the `Compile` target.  This is to replace any custom view
references with the correct `{md5}.View` style references, as we
replace the normal readable namespace with an md5 hash.

The problem was that `<ConvertResourcesCases/>` is doing a TON of
work it doesn't really need to do the second time around. For example
checking if it needs to lower case names of items.  The second pass
really only needs to worry about custom views.  In addition to that
it was also re-scanning ALL the files again.

This commit introduces a new `<ConvertCustomView/>` task.  The sole
purpose of this task is to fix up the layout files which contain
custom views.  It does nothing else.  To make this even quicker we
modify `<ConvertResourcesCases/>` to emit a mapping file,
`customview-map.txt`.  This file contains entries like:

	MonoDroid.Example.MyLayout;/fullpath/to/file.xml
	android.support.v7.widget.ActionBarOverlayLayout;/fullpath/to/some/other/file.xml

This allows us to know which files contain ANY layout.

This information in conjunction with the `acw_map.txt` file will
allow us to do a targeted update: we go through the `acw_map.txt`
values and fix up those files where we have entires in 
`customview-map.txt`.

This reduces the amount of time spent processing files quite a bit.
For a Blank Xamarin Forms app from a clean build.

	2639 ms  ConvertResourcesCases                      1 calls
	   3 ms  ConvertCustomView                          1 calls

Normally the `ConvertResourcesCases` would be called twice and would
take a total of 5-6 seconds.