[Xamarin.Android.Build.Tasks] `_CreateAapt2VersionCache` slow due to wildcard usage (#2124)
Fixes: https://github.com/xamarin/xamarin-android/issues/2121

There were a number of performance issues with the
`_CreateAapt2VersionCache` target.

Firstly the wildcards were processing the *entire*
`$(IntermediateOutputPath)` tree!  What they should have been doing
was targeting specific directories, e.g. the root of
`$(IntermediateOutputPath)` and the directories under
`$(IntermediateOutputPath)\lp`.

Secondly the target did not have a `Condition` to stop it from
running if the versions matched, so that has been added.

Thirdly, even if a target is NOT run, MSBuild will still evaluate the
`<PropertyGroup/>`s and `<ItemGroup/>`s within the Target, so we need
to add a `Conditon` on the `@(_CompiledFlataArchive)` and
`@(_CompiledFlataStamp)` items as well to prevent evaluation and
corresponding directory traversal.

With these in place the time when this target is reduced to 1ms.