[Tests] Make sure emulator is killed after tests fail (#2167)
We currently fail to kill the emulator if any of the steps executing APK tests
fail. This is because the failed task stops execution of all the other targets
defined in that file. In effect we leave running instances of the emulator on
the build bots, should such a failure happen.

This diff adds `ContinueOnError="ErrorAndContinue" to each task in the target
file so that we are sure to hit the `ReleaseAndroidTarget` target no matter what.