[Xamarin.Android.Build.Tasks] combine CheckTargetFrameworks/ResolveAssemblies
Context: https://gist.github.com/jonathanpeppers/a207a8828317460239a12a8b0cc918d7

When reviewing how many times we are using Mono.Cecil to open every
assembly, it was happening 9 times!

Unfortunately it is not as easy as just "caching" something, the
simplest approach for now is to start combining functionality that are
in multiple tasks right now.

A simple first example is the `CheckTargetFrameworks` MSBuild task,
which runs after the `Compile` and `ResolveAssemblies` MSBuild tasks.
`CheckTargetFrameworks` has the job of emitting warnings about
`[assembly: System.Runtime.Versioning.TargetFramework]`. We can easily
do this work in `ResolveAssemblies` as it is already looping over all
the assemblies' custom attributes for other reasons.

Changes to make this happen:
- Removed `CheckTargetFrameworks` and the `_CheckTargetFrameworks`
  target, adding the required `TargetFrameworkVersion` and
  `ProjectFile` properties to `ResolveAssemblies`.
- Added a new `CheckAssemblyAttributes` where we can add to a `switch`
  statement for new attribute types in the future as needed.
- Added a dictionary of `api_levels`, and copied over the logic from
  `CheckTargetFrameworks`

Before:

    105 ms  CheckTargetFrameworks                      1 calls
    188 ms  ResolveAssemblies                          1 calls

After:

    216 ms  ResolveAssemblies                          1 calls

I timed the `tests/Xamarin.Forms-Performance-Integration` project in
this repository: a build with no changes.

Since `CheckTargetFrameworks` is removed completely, we can avoid the
time it was taking to open every assembly again.

This saves around 77ms.